ajuste type float

+++ fonte/Float.java
@@ -2,7 +2,7 @@ public class Float extends Variable {
     public float valor;
 
     public Float(String name, float valor){
-        this.type = "int";
+        this.type = "float";
         this.name = name;
         this.valor = valor;
     }
